requirements.txt
langchain-groq
langchain-community
langchain-tavily
python-dotenv
uvicorn
fastapi
pydantic
streamlit
langgraph
langchain-core
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0